Change handler name, better targeting for the value update, and add comments/docs
